+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.12.9] - in progress
+
+- New function ``streamer.tee``.
+- New class ``streamer.IterProcessQueue``.
+
+
 ## [0.12.8] - 2023-05-17
 
 - Removed ``mpserver.{ProcessWorker, ThreadWorker}``; just use ``Worker``.